Add work experience highlights
1. Add support for work experience highlights.
2. Add a wonky conversion of the Handlebars context when `this` is used
   in iteration of `highlights` with the `markdown` helper.

Elaborating on point 2, when `#each highlights` passes `this` as a value
to the `markdown` helper function, the `context` is turned into an
object where each key of the object is a string containing the number
`0` through the length of the string contained within the `highlights`
array, and the value is each character of the string, such as:

```json
{
  '0': 'H',
  '1': 'i',
  '2': 'g',
  '3': 'h',
  '4': 'l',
  '5': 'i',
  '6': 'g',
  '7': 'h',
  '8': 't',
}
```

The added conversion joins the values of the above object into the
string `"Highlight"`. This is either a bug in `handlebars-utils`, how
`handlebars-utils` is used in this theme or how I've authored the code
that iterates over `highlights`. Stupid, but it works.